+++ clickhouse-cli-client/src/main/java/com/clickhouse/client/cli/ClickHouseCommandLine.java
@@ -41,7 +41,7 @@ import com.clickhouse.data.ClickHousePassThruStream;
 import com.clickhouse.data.ClickHouseUtils;
 import com.clickhouse.logging.Logger;
 import com.clickhouse.logging.LoggerFactory;
-
+@Deprecated (since = "0.6.0", forRemoval = true)
 public class ClickHouseCommandLine implements AutoCloseable {
     private static final Logger log = LoggerFactory.getLogger(ClickHouseCommandLine.class);